route products POST
