+++ api/controllers/products/postProducts.js
@@ -2,6 +2,7 @@ const createHttpError = require('http-errors');
 const { Product, Category } = require('../../database/models');
 const { endpointResponse } = require('../../helpers/success');
 const { ErrorObject } = require('../../helpers/error');
+const { get } = require('../../routes');
 
 module.exports = {
     postProducts: async (req, res, next) => {
@@ -11,7 +12,6 @@ module.exports = {
             image,
             price,
             categories} = req.body;
-
             try {
                 const createProducts = await Product.create({
                     name,
@@ -20,13 +20,27 @@ module.exports = {
                     price,        
                 });
 
-                Category.addCategory(createProducts);
+                const dbCategory = await Category.findall({
+                    where: {
+                        name: categories
+                    }
+                });
+
+             
+                Category.addCategory(dbCategory);
+                endpointResponse({
+                    res,
+                    code: 200,
+                    message: 'Test retrieved successfully',
+                    body: createProducts,
+                  });
                 
             } catch (error) {
-                
+                    const httpError = createHttpError(
+                        error.statusCode,
+                        `[Error retrieving products] - [product - POST]: ${error.message}`,
+                      )
+                      next(httpError)          
+                }   
             }
-    }
-
-        
-
 };
