+++ api/routes/index.js
@@ -1,10 +1,10 @@
 const express = require('express')
-const { getProducts } = require('../controllers/getProducts/getProducts')
+const { postProducts } = require('../controllers/products/postProducts')
 const router = express.Router()
 
 // example of a route with index controller get function
 //router.get('/', get)
-router.get('/products', getProducts)
+router.post('/products', postProducts)
 
 
 module.exports = router
